requirements.txt
# C3-Fuse Requirements
# Python 3.8+

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Point cloud processing
open3d>=0.17.0
# MinkowskiEngine>=0.5.4  # Uncomment for sparse convolutions (requires manual install)

# Computer vision
opencv-python>=4.7.0
opencv-contrib-python>=4.7.0

# Scientific computing
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.2.0
pandas>=1.5.0

# 3D geometry
trimesh>=3.20.0
pymeshlab>=2022.2  # For mesh processing

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0
# mpl-toolkits-basemap>=1.3.0  # Deprecated, use cartopy instead if needed

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Progress bars
tqdm>=4.64.0

# Logging and monitoring
tensorboard>=2.12.0
wandb>=0.14.0  # Optional: for experiment tracking

# Image processing
pillow>=9.4.0
imageio>=2.26.0

# Statistical analysis
statsmodels>=0.13.5

# PDF generation (for reports)
reportlab>=3.6.12
matplotlib>=3.6.0

# Point cloud registration
# pyicp>=0.1.0  # Uncomment if using Python ICP wrapper

# Miscellaneous
h5py>=3.8.0
pytransform3d>=2.4.0

# Testing (optional)
pytest>=7.2.0
pytest-cov>=4.0.0

# Code quality (optional)
black>=23.1.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0
